# Conflicts:
#	src/com/company/Magatzem.java

+++ src/com/company/AbstractProduct.java
@@ -1,5 +1,5 @@
 package com.company;
 
 public abstract class AbstractProduct {
-    public abstract int ObtenirPreu();
+    public abstract double ObtenirPreu();
 }
